+++ deltachat/rpc_test.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/chatmail/rpc-client-go/deltachat/option"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -36,10 +35,10 @@ func TestRpc_SetChatVisibility(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		chatId, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityPinned))
-		assert.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityArchived))
-		assert.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityNormal))
+		require.Nil(t, err)
+		require.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityPinned))
+		require.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityArchived))
+		require.Nil(t, rpc.SetChatVisibility(accId, chatId, ChatVisibilityNormal))
 	})
 }
 
@@ -58,8 +57,8 @@ func TestAccount_Select(t *testing.T) {
 	t.Parallel()
 	acfactory.WithRpc(func(rpc *Rpc) {
 		accId, err := rpc.AddAccount()
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.SelectAccount(accId))
+		require.Nil(t, err)
+		require.Nil(t, rpc.SelectAccount(accId))
 	})
 }
 
@@ -67,8 +66,8 @@ func TestAccount_StartIo(t *testing.T) {
 	t.Parallel()
 	acfactory.WithRpc(func(rpc *Rpc) {
 		accId, err := rpc.AddAccount()
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.StartIo(accId))
+		require.Nil(t, err)
+		require.Nil(t, rpc.StartIo(accId))
 	})
 }
 
@@ -76,8 +75,8 @@ func TestAccount_StopIo(t *testing.T) {
 	t.Parallel()
 	acfactory.WithRpc(func(rpc *Rpc) {
 		accId, err := rpc.AddAccount()
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.StopIo(accId))
+		require.Nil(t, err)
+		require.Nil(t, rpc.StopIo(accId))
 	})
 }
 
@@ -103,8 +102,8 @@ func TestAccount_Info(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
 		info, err := rpc.GetInfo(accId)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, info["sqlite_version"])
+		require.Nil(t, err)
+		require.NotEmpty(t, info["sqlite_version"])
 	})
 }
 
@@ -112,8 +111,8 @@ func TestAccount_Size(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
 		size, err := rpc.GetAccountFileSize(accId)
-		assert.Nil(t, err)
-		assert.NotEqual(t, size, 0)
+		require.Nil(t, err)
+		require.NotEqual(t, size, 0)
 	})
 }
 
@@ -121,49 +120,49 @@ func TestAccount_IsConfigured(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
 		configured, err := rpc.IsConfigured(accId)
-		assert.Nil(t, err)
-		assert.False(t, configured)
+		require.Nil(t, err)
+		require.False(t, configured)
 
-		assert.Nil(t, rpc.Configure(accId))
+		require.Nil(t, rpc.Configure(accId))
 
 		configured, err = rpc.IsConfigured(accId)
-		assert.Nil(t, err)
-		assert.True(t, configured)
+		require.Nil(t, err)
+		require.True(t, configured)
 	})
 }
 
 func TestAccount_SetConfig(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
-		assert.Nil(t, rpc.SetConfig(accId, "displayname", option.Some("test name")))
+		require.Nil(t, rpc.SetConfig(accId, "displayname", option.Some("test name")))
 		name, err := rpc.GetConfig(accId, "displayname")
-		assert.Nil(t, err)
-		assert.Equal(t, name.Unwrap(), "test name")
+		require.Nil(t, err)
+		require.Equal(t, name.Unwrap(), "test name")
 
 		err = rpc.BatchSetConfig(accId, map[string]option.Option[string]{
 			"displayname": option.Some("new name"),
 			"selfstatus":  option.Some("test status"),
 		})
-		assert.Nil(t, err)
+		require.Nil(t, err)
 		name, err = rpc.GetConfig(accId, "displayname")
-		assert.Nil(t, err)
-		assert.Equal(t, name.Unwrap(), "new name")
+		require.Nil(t, err)
+		require.Equal(t, name.Unwrap(), "new name")
 
-		assert.Nil(t, rpc.SetConfig(accId, "selfavatar", option.Some(acfactory.TestImage())))
+		require.Nil(t, rpc.SetConfig(accId, "selfavatar", option.Some(acfactory.TestImage())))
 	})
 }
 
 func TestAccount_Remove(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
-		assert.Nil(t, rpc.RemoveAccount(accId))
+		require.Nil(t, rpc.RemoveAccount(accId))
 	})
 }
 
 func TestAccount_Configure(t *testing.T) {
 	t.Parallel()
 	acfactory.WithUnconfiguredAccount(func(rpc *Rpc, accId AccountId) {
-		assert.Nil(t, rpc.Configure(accId))
+		require.Nil(t, rpc.Configure(accId))
 	})
 }
 
@@ -171,12 +170,12 @@ func TestAccount_Contacts(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		ids, err := rpc.GetContactIds(accId, 0, option.None[string]())
-		assert.Nil(t, err)
-		assert.Empty(t, ids)
+		require.Nil(t, err)
+		require.Empty(t, ids)
 
 		ids, err = rpc.GetContactIds(accId, 0, option.Some("unknown"))
-		assert.Nil(t, err)
-		assert.Empty(t, ids)
+		require.Nil(t, err)
+		require.Empty(t, ids)
 	})
 }
 
@@ -184,17 +183,17 @@ func TestAccount_GetContactByAddr(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		contactId, err := rpc.CreateContact(accId, "null@localhost", "test")
-		assert.Nil(t, err)
-		assert.NotNil(t, contactId)
+		require.Nil(t, err)
+		require.NotNil(t, contactId)
 
 		contactId2, err := rpc.LookupContactIdByAddr(accId, "unknown@example.com")
-		assert.Nil(t, err)
-		assert.True(t, contactId2.IsNone())
+		require.Nil(t, err)
+		require.True(t, contactId2.IsNone())
 
 		contactId2, err = rpc.LookupContactIdByAddr(accId, "null@localhost")
-		assert.Nil(t, err)
-		assert.True(t, contactId2.IsSome())
-		assert.Equal(t, contactId, contactId2.Unwrap())
+		require.Nil(t, err)
+		require.True(t, contactId2.IsSome())
+		require.Equal(t, contactId, contactId2.Unwrap())
 	})
 }
 
@@ -202,17 +201,17 @@ func TestAccount_BlockedContacts(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		contactId, err := rpc.CreateContact(accId, "null@localhost", "test")
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		blocked, err := rpc.GetBlockedContacts(accId)
-		assert.Nil(t, err)
-		assert.Empty(t, blocked)
+		require.Nil(t, err)
+		require.Empty(t, blocked)
 
-		assert.Nil(t, rpc.BlockContact(accId, contactId))
+		require.Nil(t, rpc.BlockContact(accId, contactId))
 
 		blocked, err = rpc.GetBlockedContacts(accId)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, blocked)
+		require.Nil(t, err)
+		require.NotEmpty(t, blocked)
 	})
 }
 
@@ -220,7 +219,7 @@ func TestAccount_CreateBroadcastList(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		_, err := rpc.CreateBroadcastList(accId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 	})
 }
 
@@ -228,22 +227,37 @@ func TestAccount_CreateGroup(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		_, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 	})
 }
 
-func TestAccount_QrCode(t *testing.T) {
+func TestAccount_GetChatSecurejoinQrCodeSvg(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc1 *Rpc, accId1 AccountId) {
 		qrdata, svg, err := rpc1.GetChatSecurejoinQrCodeSvg(accId1, option.None[ChatId]())
-		assert.Nil(t, err)
-		assert.NotEmpty(t, qrdata)
-		assert.NotEmpty(t, svg)
+		require.Nil(t, err)
+		require.NotEmpty(t, qrdata)
+		require.NotEmpty(t, svg)
 
 		acfactory.WithOnlineAccount(func(rpc2 *Rpc, accId2 AccountId) {
-			acfactory.IntroduceEachOther(rpc1, accId1, rpc2, accId2)
 			_, err := rpc2.SecureJoin(accId2, qrdata)
-			assert.Nil(t, err)
+			require.Nil(t, err)
+			acfactory.WaitForEvent(rpc1, accId1, EventSecurejoinInviterProgress{})
+			acfactory.WaitForEvent(rpc2, accId2, EventSecurejoinJoinerProgress{})
+		})
+	})
+}
+
+func TestAccount_GetChatSecurejoinQrCode(t *testing.T) {
+	t.Parallel()
+	acfactory.WithOnlineAccount(func(rpc1 *Rpc, accId1 AccountId) {
+		qrdata, err := rpc1.GetChatSecurejoinQrCode(accId1, option.None[ChatId]())
+		require.Nil(t, err)
+		require.NotEmpty(t, qrdata)
+
+		acfactory.WithOnlineAccount(func(rpc2 *Rpc, accId2 AccountId) {
+			_, err := rpc2.SecureJoin(accId2, qrdata)
+			require.Nil(t, err)
 			acfactory.WaitForEvent(rpc1, accId1, EventSecurejoinInviterProgress{})
 			acfactory.WaitForEvent(rpc2, accId2, EventSecurejoinJoinerProgress{})
 		})
@@ -256,20 +270,20 @@ func TestAccount_ImportBackup(t *testing.T) {
 	passphrase := option.Some("password")
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		dir := acfactory.MkdirTemp()
-		assert.Nil(t, rpc.ExportBackup(accId, dir, passphrase))
+		require.Nil(t, rpc.ExportBackup(accId, dir, passphrase))
 		files, err := os.ReadDir(dir)
-		assert.Nil(t, err)
-		assert.Equal(t, len(files), 1)
+		require.Nil(t, err)
+		require.Equal(t, len(files), 1)
 		backup = filepath.Join(dir, files[0].Name())
-		assert.FileExists(t, backup)
+		require.FileExists(t, backup)
 	})
 
 	acfactory.WithRpc(func(rpc *Rpc) {
 		accId, err := rpc.AddAccount()
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.ImportBackup(accId, backup, passphrase))
+		require.Nil(t, err)
+		require.Nil(t, rpc.ImportBackup(accId, backup, passphrase))
 		_, err = rpc.GetSystemInfo()
-		assert.Nil(t, err)
+		require.Nil(t, err)
 	})
 }
 
@@ -277,17 +291,17 @@ func TestAccount_ExportBackup(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		dir := acfactory.MkdirTemp()
-		assert.Nil(t, rpc.ExportBackup(accId, dir, option.Some("test-phrase")))
+		require.Nil(t, rpc.ExportBackup(accId, dir, option.Some("test-phrase")))
 		files, err := os.ReadDir(dir)
-		assert.Nil(t, err)
-		assert.Equal(t, len(files), 1)
+		require.Nil(t, err)
+		require.Equal(t, len(files), 1)
 	})
 }
 
 func TestAccount_GetBackup(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc1 *Rpc, accId1 AccountId) {
-		go func() { assert.Nil(t, rpc1.ProvideBackup(accId1)) }()
+		go func() { require.Nil(t, rpc1.ProvideBackup(accId1)) }()
 		var err error
 		var qrData string
 		qrData, err = rpc1.GetBackupQr(accId1)
@@ -295,16 +309,16 @@ func TestAccount_GetBackup(t *testing.T) {
 			time.Sleep(time.Millisecond * 200)
 			qrData, err = rpc1.GetBackupQr(accId1)
 		}
-		assert.NotNil(t, qrData)
+		require.NotNil(t, qrData)
 
 		qrSvg, err := rpc1.GetBackupQrSvg(accId1)
-		assert.Nil(t, err)
-		assert.NotNil(t, qrSvg)
+		require.Nil(t, err)
+		require.NotNil(t, qrSvg)
 
 		acfactory.WithRpc(func(rpc2 *Rpc) {
 			accId2, err := rpc2.AddAccount()
-			assert.Nil(t, err)
-			assert.Nil(t, rpc2.GetBackup(accId2, qrData))
+			require.Nil(t, err)
+			require.Nil(t, rpc2.GetBackup(accId2, qrData))
 		})
 	})
 }
@@ -313,8 +327,8 @@ func TestAccount_InitiateAutocryptKeyTransfer(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		code, err := rpc.InitiateAutocryptKeyTransfer(accId)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, code)
+		require.Nil(t, err)
+		require.NotEmpty(t, code)
 	})
 }
 
@@ -324,19 +338,19 @@ func TestAccount_FreshMsgs(t *testing.T) {
 		acfactory.WithOnlineAccount(func(rpc2 *Rpc, accId2 AccountId) {
 			chatId2 := acfactory.CreateChat(rpc2, accId2, rpc1, accId1)
 			_, err := rpc2.MiscSendTextMessage(accId2, chatId2, "hi")
-			assert.Nil(t, err)
+			require.Nil(t, err)
 			msg := acfactory.NextMsg(rpc1, accId1)
-			assert.Equal(t, msg.Text, "hi")
+			require.Equal(t, msg.Text, "hi")
 
 			msgs, err := rpc1.GetFreshMsgs(accId1)
-			assert.Nil(t, err)
-			assert.NotEmpty(t, msgs)
+			require.Nil(t, err)
+			require.NotEmpty(t, msgs)
 
-			assert.Nil(t, rpc1.MarkseenMsgs(accId1, msgs))
+			require.Nil(t, rpc1.MarkseenMsgs(accId1, msgs))
 
 			msgs, err = rpc1.GetFreshMsgs(accId1)
-			assert.Nil(t, err)
-			assert.Empty(t, msgs)
+			require.Nil(t, err)
+			require.Empty(t, msgs)
 		})
 	})
 }
@@ -345,18 +359,18 @@ func TestAccount_GetNextMsgs(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineBot(func(bot *Bot, botAcc AccountId) {
 		msgs, err := bot.Rpc.GetNextMsgs(botAcc)
-		assert.Nil(t, err)
-		assert.Empty(t, msgs)
+		require.Nil(t, err)
+		require.Empty(t, msgs)
 		acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 			msgs, err := rpc.GetNextMsgs(accId)
-			assert.Nil(t, err)
-			assert.NotEmpty(t, msgs) // messages from device chat
+			require.Nil(t, err)
+			require.NotEmpty(t, msgs) // messages from device chat
 
-			assert.Nil(t, rpc.MarkseenMsgs(accId, msgs))
+			require.Nil(t, rpc.MarkseenMsgs(accId, msgs))
 
 			msgs, err = rpc.GetNextMsgs(accId)
-			assert.Nil(t, err)
-			assert.Empty(t, msgs)
+			require.Nil(t, err)
+			require.Empty(t, msgs)
 		})
 	})
 }
@@ -365,19 +379,19 @@ func TestAccount_DeleteMsgs(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		chatId, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 		_, err = rpc.MiscSendTextMessage(accId, chatId, "hi")
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		msgs, err := rpc.GetMessageIds(accId, chatId, false, false)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, msgs)
+		require.Nil(t, err)
+		require.NotEmpty(t, msgs)
 
-		assert.Nil(t, rpc.DeleteMessages(accId, msgs))
+		require.Nil(t, rpc.DeleteMessages(accId, msgs))
 
 		msgs, err = rpc.GetMessageIds(accId, chatId, false, false)
-		assert.Nil(t, err)
-		assert.Empty(t, msgs)
+		require.Nil(t, err)
+		require.Empty(t, msgs)
 	})
 }
 
@@ -385,14 +399,14 @@ func TestAccount_SearchMessages(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		chatId, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 		msgId, err := rpc.MiscSendTextMessage(accId, chatId, "hi")
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		msgs, err := rpc.SearchMessages(accId, "hi", option.None[ChatId]())
-		assert.Nil(t, err)
-		assert.NotEmpty(t, msgs)
-		assert.Equal(t, msgId, msgs[0])
+		require.Nil(t, err)
+		require.NotEmpty(t, msgs)
+		require.Equal(t, msgId, msgs[0])
 	})
 }
 
@@ -400,21 +414,21 @@ func TestAccount_GetChatlistEntries(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		_, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		noFlag := option.None[uint]()
 		noContact := option.None[ContactId]()
 		entries, err := rpc.GetChatlistEntries(accId, noFlag, option.Some("unknown"), noContact)
-		assert.Nil(t, err)
-		assert.Empty(t, entries)
+		require.Nil(t, err)
+		require.Empty(t, entries)
 
 		entries, err = rpc.GetChatlistEntries(accId, noFlag, option.None[string](), noContact)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, entries)
+		require.Nil(t, err)
+		require.NotEmpty(t, entries)
 
 		items, err := rpc.GetChatlistItemsByEntries(accId, entries)
-		assert.Nil(t, err)
-		assert.NotEmpty(t, items)
+		require.Nil(t, err)
+		require.NotEmpty(t, items)
 	})
 }
 
@@ -433,23 +447,23 @@ func TestChat_Basics(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		chatId, err := rpc.CreateGroupChat(accId, "test group", true)
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.AcceptChat(accId, chatId))
-		assert.Nil(t, rpc.MarknoticedChat(accId, chatId))
+		require.Nil(t, err)
+		require.Nil(t, rpc.AcceptChat(accId, chatId))
+		require.Nil(t, rpc.MarknoticedChat(accId, chatId))
 		_, err = rpc.GetFirstUnreadMessageOfChat(accId, chatId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		_, err = rpc.GetBasicChatInfo(accId, chatId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		_, err = rpc.GetFullChatById(accId, chatId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
-		assert.Nil(t, rpc.BlockChat(accId, chatId))
+		require.Nil(t, rpc.BlockChat(accId, chatId))
 
 		chatId, err = rpc.CreateGroupChat(accId, "test group 2", true)
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.DeleteChat(accId, chatId))
+		require.Nil(t, err)
+		require.Nil(t, rpc.DeleteChat(accId, chatId))
 	})
 }
 
@@ -457,21 +471,21 @@ func TestChat_Groups(t *testing.T) {
 	t.Parallel()
 	acfactory.WithOnlineAccount(func(rpc *Rpc, accId AccountId) {
 		chatId, err := rpc.CreateGroupChat(accId, "test group", false)
-		assert.Nil(t, err)
-		assert.Nil(t, rpc.SetChatProfileImage(accId, chatId, option.Some(acfactory.TestImage())))
-		assert.Nil(t, rpc.SetChatProfileImage(accId, chatId, option.None[string]()))
-		assert.Nil(t, rpc.SetChatName(accId, chatId, "new name"))
+		require.Nil(t, err)
+		require.Nil(t, rpc.SetChatProfileImage(accId, chatId, option.Some(acfactory.TestImage())))
+		require.Nil(t, rpc.SetChatProfileImage(accId, chatId, option.None[string]()))
+		require.Nil(t, rpc.SetChatName(accId, chatId, "new name"))
 
 		_, err = rpc.GetChatContacts(accId, chatId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
-		assert.Nil(t, rpc.SetChatEphemeralTimer(accId, chatId, 9000))
+		require.Nil(t, rpc.SetChatEphemeralTimer(accId, chatId, 9000))
 
 		_, err = rpc.GetChatEncryptionInfo(accId, chatId)
-		assert.Nil(t, err)
+		require.Nil(t, err)
 
 		_, err = rpc.SendMsg(accId, chatId, MsgData{Text: "test message"})
-		assert.Nil(t, err)
+		require.Nil(t, err)
 	})
 }